Enable type deduction for `cast_truncate`.

+++ trapping_test.cc
@@ -111,45 +111,45 @@ void TestCastTruncate() {
   // Expect narrowing casts out of range of `R` to truncate:
   {
     i8 result;
-    EXPECT((cast_truncate<i16, i8>(i16_max, &result)));
-    EXPECT((cast_truncate<i32, i8>(i16_max, &result)));
-    EXPECT((cast_truncate<i64, i8>(i16_max, &result)));
+    EXPECT((cast_truncate<i8, i16>(i16_max, &result)));
+    EXPECT((cast_truncate<i8, i32>(i16_max, &result)));
+    EXPECT((cast_truncate<i8, i64>(i16_max, &result)));
   }
   {
     u8 result;
-    EXPECT((cast_truncate<i16, u8>(i16_max, &result)));
-    EXPECT((cast_truncate<i32, u8>(i16_max, &result)));
-    EXPECT((cast_truncate<i64, u8>(i16_max, &result)));
+    EXPECT((cast_truncate<u8, i16>(i16_max, &result)));
+    EXPECT((cast_truncate<u8, i32>(i16_max, &result)));
+    EXPECT((cast_truncate<u8, i64>(i16_max, &result)));
   }
 
   // Expect narrowing casts in range of `R` not to truncate:
   {
     i8 result;
-    EXPECT(!(cast_truncate<i16, i8>(i8_max, &result)));
-    EXPECT(!(cast_truncate<i32, i8>(i8_max, &result)));
-    EXPECT(!(cast_truncate<i64, i8>(i8_max, &result)));
+    EXPECT(!(cast_truncate<i8, i16>(i8_max, &result)));
+    EXPECT(!(cast_truncate<i8, i32>(i8_max, &result)));
+    EXPECT(!(cast_truncate<i8, i64>(i8_max, &result)));
   }
   {
     u8 result;
-    EXPECT(!(cast_truncate<i16, u8>(i8_max, &result)));
-    EXPECT(!(cast_truncate<i32, u8>(i8_max, &result)));
-    EXPECT(!(cast_truncate<i64, u8>(i8_max, &result)));
+    EXPECT(!(cast_truncate<u8, i16>(i8_max, &result)));
+    EXPECT(!(cast_truncate<u8, i32>(i8_max, &result)));
+    EXPECT(!(cast_truncate<u8, i64>(i8_max, &result)));
   }
 
   // Expect negative values to truncate when cast to unsigned:
   {
     u32 result;
-    EXPECT((cast_truncate<i16, u32>(-1, &result)));
-    EXPECT((cast_truncate<i32, u32>(-1, &result)));
-    EXPECT((cast_truncate<i64, u32>(-1, &result)));
+    EXPECT((cast_truncate<u32, i16>(-1, &result)));
+    EXPECT((cast_truncate<u32, i32>(-1, &result)));
+    EXPECT((cast_truncate<u32, i64>(-1, &result)));
   }
 
   // Expect positive, in-range values to truncate when cast to unsigned:
   {
     u32 result;
-    EXPECT(!(cast_truncate<i16, u32>(i16_max, &result)));
-    EXPECT(!(cast_truncate<i32, u32>(i16_max, &result)));
-    EXPECT(!(cast_truncate<i64, u32>(i16_max, &result)));
+    EXPECT(!(cast_truncate<u32, i16>(i16_max, &result)));
+    EXPECT(!(cast_truncate<u32, i32>(i16_max, &result)));
+    EXPECT(!(cast_truncate<u32, i64>(i16_max, &result)));
   }
 }
